trading-aggregator/pyproject.toml
[tool.poetry]
name = "trading-aggregator"
version = "0.1.0"
description = ""
authors = []
license = "Proprietary"


[tool.poetry.dependencies]
python = "^3.10"
websockets = "^11.0.3"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
